Changed profile URL for JasonB
I'm either to stupid or it's not possible to add the profile page in a Hugo site as is. Every attempt I made messed with my site theme and the profile theme. 
I hope this is a decent compromise.

Thanks to Tymscar for being patient with me as I was setting up DNS properly so HTTPS would work.